pulling out utility methods
