+++ app/controllers/paging_schedule_controller.rb
@@ -4,6 +4,8 @@
 #  Controller for returning the paging schedule for async requests
 ###
 class PagingScheduleController < ApplicationController
+  include Folio::TypesUtils
+
   rescue_from PagingSchedule::ScheduleNotFound do
     render plain: 'Schedule not found', status: :not_found
   end
@@ -52,35 +54,5 @@ class PagingScheduleController < ApplicationController
       origin: params[:origin],
       destination:
     )
-  end
-
-  # For FOLIO, destination is specified as service point
-  # Convert service point to library for scheduling and library hours
-  def map_to_library(service_point_code)
-    return service_point_code if service_point_code == 'SCAN'
-
-    # If the service point does not match, default to GREEN-LOAN
-    code = valid_service_point_code(service_point_code) ? service_point_code : Settings.folio.default_service_point
-    service_point_id = get_service_point_id(code)
-    library_id = get_library_for_service_point(service_point_id)
-
-    # Find the library code associated with this library id
-    Folio::Types.instance.get_type('libraries').find { |library| library['id'] == library_id }['code']
-  end
-
-  # Find the service point ID based on this service point code
-  def get_service_point_id(service_point_code)
-    Folio::Types.instance.service_points.values.find { |v| v.code == service_point_code }&.id
-  end
-
-  # Find the library id for the location with which this service point is associated
-  def get_library_for_service_point(service_point_id)
-    loc = Folio::Types.instance.get_type('locations').find { |location| location['primaryServicePoint'] == service_point_id }
-    loc.present? && loc.key?('libraryId') ? loc['libraryId'] : nil
-  end
-
-  # Check if valid service point, otherwise default to GREEN
-  def valid_service_point_code(service_point_code)
-    Folio::Types.instance.service_points.values.find { |v| v.code == service_point_code }.present?
   end
 end
